Fix fetching of images from tumblr
When fetching a tumblr image, tumblr will normally return a proxy html
page to contain the image.  From within that page it then fetches the
image again where the referer is set to the image url, causing the
actual image to be retrieved instead of the proxy page.

This change performs a HTTP GET request with the Referer set so that the
actual image is retrieved.

Note that users that have used the previous implementation will have a
cache full of images that are actually html pages, causing TAI to render
those cached images as a black image.  It is advisable for the users to
clear their cache.