Refactor ShowImageFunction to reduce complexity
The new base class, CustomFunctionExtended, uses reflection to determine
the appropriate onCall(...) method in the derived class to be invoked
based on the argument types passed to call().  This greatly simplifies
the implementation in the derived class that needs only to provide the
various onCall(...) methods with the required parameter types.

When the base class is unable to match an appropriate onCall(...) method
in the derived class because there is no parameter type match, log
information is generated to support the developer of the personality.
For example if there is a call to showImage(10, "myImage.jpg"), the
following log is produced:
"
  06:54:28 pm SEVERE: No match for function call showImage(Integer, String)
  06:54:28 pm INFO: Candidate functions are:
  06:54:28 pm INFO:     showImage(File)
  06:54:28 pm INFO:     showImage(File, Integer)
  06:54:28 pm INFO:     showImage(MediaURL)
  06:54:28 pm INFO:     showImage(MediaURL, Integer)
  06:54:28 pm INFO:     showImage(String)
  06:54:28 pm INFO:     showImage(String, Integer)
"

CustomFunctionExtended derives from the existing CustomFunction so that
all of the existing functions that derive from CustomFunction are
unaffected by this change.  If desired, the other functions could also
be modified in the same way, one at a time if needed.  If all of the
functions are eventually modified to derive from CustomFunctionExtended,
then CustomFunction and CustomFunctionExtended could be merged together.