Refactor SendMessageFunction function to use CustomFunctionExtended
The previous implementation of sendMessage(String, Number, Boolean) didn't
multiply the duration value by 1000 to convert from seconds to
milliseconds.  It's not clear whether this was intentional to provide the
possibility of passing -1 to customMessage() as the duration which has a
special interpretation.  However, neither the Spicy or Mischevious
personalites used this function, and it's not mentioned on the Wiki page.

Tested with the following script:
"
lockImages();

// sendMessage(String)
sendMessage('Mary had a little lamb');

// sendMessage(String, Number)
sendMessage('its fleece was white as snow', 2);
sendMessage('and everywhere that Mary went', 3.5);

// sendMessage(String, String)
sendMessage('Mary had a little lamb', 'Images/Spicy/Assistant/1/124259409_004_122_372lo.jpg');

// sendMessage(String, Number, Boolean)
sendMessage('who\s fleece was white as snow', 10, true);
sendMessage('and everywhere that Mary went', 3.5, false);

sleep(5);

// Bad calls
sendMessage();
sendMessage(null);
sendMessage(2, 'Mary had a little lamb');
sendMessage('its fleece was white as snow', 2, 3.5);
"