Fortify against media stability issues
Tested with:
    git@github.com:FrozenWolf4887/TeaseAIJava-TestPersonality.git

This separates the media handlers into three respective modules, video,
audio, and image.  Additionally it addresses various issues found with
testing, although most of the issues only pertain to error conditions.

Video playback issues addressed:
* Disposes of MediaPlayers when finished or failed
* Fix hanging when playing invalid files
* Fix images remaining locked when playing invalid files
* Fix crash calling stopVideo() when no video is playing
* Stop previously playing video before starting a new one

Audio playback issues addressed:
* Disposes of MediaPlayers when finished or failed
* Fix hanging when playing invalid files
* Fix images being unlocked when audio has finished playing
* Fix stopAllAudio() only stopping the last playing occurrence of a file
* Cleans up list of playing audio

There is an outstanding issue that was also present in TAJ 1.3 where
calling playVideo('http://...') would often hang.  The contained
MediaPlayer object fails to issue any callbacks when the video has
finished, and its status remains as PLAYING.  Calling stop() on the
MediaPlayer does nothing.  This causes a hang in TAJ when waiting for
the video to finish either successfully or in error.  This problem may
be limited to Linux and hasn't been evaluated in Windows.  It occurs
with both OpenJFX 15.0.1 and OpenJFX 16; other versions have not been
tested.